springboot

+++ src/main/java/cn/edu/guet/controller/StudentController.java
@@ -6,70 +6,58 @@ import cn.edu.guet.model.Student;
 import cn.edu.guet.service.IStudentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
-@Controller
+@RestController
 @RequestMapping("student")
 public class StudentController {
 
     @Autowired
     IStudentService studentService;
-    @Autowired
-    ResponseTemplate responseTemplate;
     @RequestMapping(value = "getStudent",method = RequestMethod.GET)
-    @ResponseBody
+
      public ResponseTemplate getStudentById(String id){
-        Student student = studentService.getStudentById(id);
-        responseTemplate.setCode(200);
-        responseTemplate.setMessage("查询成功");
-        responseTemplate.setData(student);
-        return responseTemplate;
+        if(id!=null){
+            Student student = studentService.getStudentById(id);
+            return  ResponseTemplate.result(200,"查询成功",student);
+        }
+        return  ResponseTemplate.result(200,"ID格式错误",null);
     }
 
     @RequestMapping(value = "getAllStudent",method = RequestMethod.GET)
-    @ResponseBody
     public ResponseTemplate getAllStudent(){
         List<Student> allStudent = studentService.getAllStudent();
-        responseTemplate.setData(allStudent);
-        responseTemplate.setCode(200);
-        responseTemplate.setMessage("查询成功");
-        return responseTemplate;
+        return  ResponseTemplate.result(200,"查询成功",allStudent);
     }
 
     @RequestMapping(method = RequestMethod.GET)
-    @ResponseBody
+
     public ResponseTemplate deleteStudent(String id){
         try {
-            studentService.deleteStudentById(id);
-
+            if(id == null){
+                studentService.deleteStudentById(id);
+                return ResponseTemplate.result(200,"删除成功",null);
+            }
         } catch (Exception e) {
             e.printStackTrace();
+            return ResponseTemplate.result(401,"删除失败",null);
         }
-
-        return null;
+        return ResponseTemplate.result(401,"ID为空",null);
     }
     @RequestMapping(value = "updateStudent",method = RequestMethod.POST)
-    @ResponseBody
-
     public ResponseTemplate updateStudent(@RequestBody Student student){
         try {
-            System.out.println("student.getId() = " + student.getId());
-            studentService.updateStudent(student);
-            responseTemplate.setCode(200);
-            responseTemplate.setMessage("更新成功");
-            responseTemplate.setData(student);
+            if(student!=null){
+                studentService.updateStudent(student);
+                return ResponseTemplate.result(200,"更新成功",null);
+            }
         } catch (Exception e) {
             e.printStackTrace();
-            responseTemplate.setCode(401);
-            responseTemplate.setMessage("更新失败");
-            responseTemplate.setData(student);
+            return ResponseTemplate.result(401,"更新失败",null);
         }
-        return responseTemplate;
+        return ResponseTemplate.result(401,"参数为空",null);
     }
 
 }
